+++ scripts/migration/src/transforms/createClassElementFromModuleExports.ts
